// apps/desktop/src/ipc/generated/ProjectFileDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AssetInfo } from "./AssetInfo";

export type ProjectFileDto = { id: string, displayName: string, folderId: string | null, notes: string | null, source: string | null, asset: AssetInfo, createdAt: number, updatedAt: number, rev: number, };